fix: keep stock trades when clearing swap & bridge history
