+++ packages/kit-bg/src/services/ServiceSwap.ts
@@ -40,6 +40,7 @@ import { equalsIgnoreCase } from '@onekeyhq/shared/src/utils/stringUtils';
 import {
   isPrivateSendSwapHistoryItem,
   isSamePrivateSendSwapHistoryItem,
+  isStockSwapHistoryItem,
   isSwapHistoryProtocolExcluded,
 } from '@onekeyhq/shared/src/utils/swapHistoryUtils';
 import {
@@ -2056,6 +2057,9 @@ export default class ServiceSwap extends ServiceBase {
     statuses?: ESwapTxHistoryStatus[],
     options?: {
       excludeProtocols?: EProtocolOfExchange[];
+      // Keep stock trades (the Swap/Bridge tab hides them via the token-level
+      // isStock flag, so clearing it must not delete stock orders).
+      excludeStock?: boolean;
     },
   ) {
     await this.backgroundApi.simpleDb.swapHistory.deleteSwapHistoryItem(
@@ -2075,6 +2079,9 @@ export default class ServiceSwap extends ServiceBase {
         ) {
           return false;
         }
+        if (options?.excludeStock && isStockSwapHistoryItem(item)) {
+          return false;
+        }
         return statuses ? statuses.includes(item.status) : true;
       })
       .map((item) =>
@@ -2091,6 +2098,9 @@ export default class ServiceSwap extends ServiceBase {
             excludeProtocols: options?.excludeProtocols,
           })
         ) {
+          return true;
+        }
+        if (options?.excludeStock && isStockSwapHistoryItem(item)) {
           return true;
         }
         return statuses ? !statuses.includes(item.status) : false;
